Add comprehensive integration testing and documentation
- Add 23 unit tests with MockWebServer (no external dependencies)
- Add integration test framework with TestContainers for Docker testing
- Add v4-specific integration tests for EMQX 4.4.8 compatibility
- Add docker-compose.yml for local MQTT broker testing
- Add run-integration-tests.sh helper script for easy test execution
- Add README_INTEGRATION_TESTS.md with detailed setup and usage guides
- Update pom.xml with Paho MQTT and TestContainers dependencies
- Update README.md with improved testing documentation and project structure
- All unit tests pass (23/23)
- Docker EMQX 4.4.8 confirmed working with basic auth

Tests verify:
- HTTP request/response handling
- JSON serialization/deserialization
- Error handling (auth, HTTP errors)
- API contract (paths, headers, methods)
- Full EMQX REST API surface (nodes, users, clients, sessions, subscriptions)

+++ src/main/java/com/apidech/sdk/emqxsdk/NodeDto.java
@@ -9,13 +9,13 @@ public class NodeDto {
 	public String  version;
 	public String  status;
 	public Long    uptime_ms;
-	public Integer memory_total;
-	public Integer memory_used;
-	public Integer process_memory;
+	public Long    memory_total;   // bytes
+	public Long    memory_used;    // bytes
+	public Long    process_memory; // bytes
 	public Integer max_fds;
 	public Integer threads;
 	public Integer socks_total;
-	public Integer live_connections;  // NUMBER 0/1 from EMQX API  // NUMBER 0/1 from EMQX API
+	public Integer live_connections;
 
 	// --- Builder support ---
 
@@ -30,9 +30,9 @@ public class NodeDto {
 		public Builder version(String v)           { dto.version          = v;  return this; }
 		public Builder status(String v)            { dto.status           = v;  return this; }
 		public Builder uptime_ms(Long v)           { dto.uptime_ms        = v;  return this; }
-		public Builder memory_total(Integer v)     { dto.memory_total     = v;  return this; }
-		public Builder memory_used(Integer v)      { dto.memory_used      = v;  return this; }
-		public Builder process_memory(Integer v)   { dto.process_memory   = v;  return this; }
+		public Builder memory_total(Long v)        { dto.memory_total     = v;  return this; }
+		public Builder memory_used(Long v)         { dto.memory_used      = v;  return this; }
+		public Builder process_memory(Long v)      { dto.process_memory   = v;  return this; }
 		public Builder max_fds(Integer v)          { dto.max_fds          = v;  return this; }
 		public Builder threads(Integer v)          { dto.threads          = v;  return this; }
 		public Builder socks_total(Integer v)      { dto.socks_total      = v;  return this; }
